backend/requirements.txt
langchain-openai
langchain-core
langgraph
duckduckgo-search
beautifulsoup4
python-dotenv
typer
rich
fastapi
uvicorn
requests

